% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as.data.frame.svy}
\alias{as.data.frame.svy}
\title{convert a svy to a dataframe, splitting matrices to multiple columns}
\usage{
\method{as.data.frame}{svy}(x, ...)
}
\description{
convert a svy to a dataframe, splitting matrices to multiple columns
}
