% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cols.R
\name{bind_cols.svy}
\alias{bind_cols.svy}
\title{bind `svy`s column-wise, preserving matrices and the attributes of the first `svy`}
\usage{
bind_cols.svy(l, ..., .names = getOption("preserve_atts", c("node",
  "class", "group", "languages", "groups")))
}
\arguments{
\item{l}{list of `svy`s to bind}

\item{...}{other arguments to bind_cols}

\item{.names}{names of attributes to preserve from `l[[1]]`}
}
\description{
bind `svy`s column-wise, preserving matrices and the attributes of the first `svy`
}
