% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{collapse.svq}
\alias{collapse.svq}
\title{S3 method to collapse a repeat \code{svq} into a single \code{svy}}
\usage{
\method{collapse}{svq}(x, index = 1:NROW(x))
}
\description{
internally, a \code{svq} of type \code{repeat} (roster) is just a list of
\code{svy}s with the same \code{node} attribute.  This function rbinds
those tibbles into a single survey.
}
