% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{connection}
\alias{connection}
\title{access or set the default database connection}
\usage{
connection(host = getOption("svyDBHost"),
  user = getOption("svyDBUser"), password = getOption("svyDBPassword"),
  dbname = getOption("svyDBName"), schema = getOption("svyDBSchema"))
}
\description{
access or set the default database connection
}
