% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{copy_atts}
\alias{copy_atts}
\title{copy some attributes from one object to another, recursively}
\usage{
copy_atts(to, from, atts = getOption("svyAttrIncl", c("group", "node",
  "languages")), recursive = TRUE)
}
\description{
copy some attributes from one object to another, recursively
}
