% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_format.R
\name{db_format.svq}
\alias{db_format.svq}
\title{prepare \code{svy} a data column for insertion in a database}
\usage{
\method{db_format}{svq}(q)
}
\description{
prepare \code{svy} a data column for insertion in a database
}
