% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive.R
\name{drive_download_recursive}
\alias{drive_download_recursive}
\title{wrapper for drive_download to allow recursive downloading of directories}
\usage{
drive_download_recursive(f, ...)
}
\description{
wrapper for drive_download to allow recursive downloading of directories
}
