% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%*\%}
\alias{\%*\%}
\title{compose two or more functions (mainly for use in *apply and plyr functions)}
\usage{
f \%*\% g
}
\description{
compose two or more functions (mainly for use in *apply and plyr functions)
}
