% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_curl}
\alias{kobo_curl}
\title{connect to the kobo api server and return the connection}
\usage{
kobo_curl(endpoint, dat = NULL, method = "GET", options = list(),
  headers = list(), server = getOption("koboServer"),
  token = getOption("koboToken"))
}
\description{
connect to the kobo api server and return the connection
}
