% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_data}
\alias{kobo_data}
\title{retrieve data from kobo}
\usage{
kobo_data(formid = getOption("koboID"),
  server = getOption("koboServer"), token = getOption("koboToken"),
  raw = TRUE)
}
\description{
retrieve data from kobo
}
