% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_form}
\alias{kobo_form}
\title{retrieve a form from kobo}
\usage{
kobo_form(formid = getOption("koboID"),
  server = getOption("koboServer"), token = getOption("koboToken"))
}
\description{
retrieve a form from kobo
}
