% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{kobo_submit}
\alias{kobo_submit}
\title{submit an instance to the kobo server}
\usage{
kobo_submit(file, formid = getOption("koboID"),
  server = getOption("koboServer"), token = getOption("koboToken"))
}
\description{
submit an instance to the kobo server
}
