% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{labels.svq}
\alias{labels.svq}
\title{extract the choice labels of a multiple choice question in a language}
\usage{
\method{labels}{svq}(x, lang = if (is.null(languages(x))) NULL else
  getOption("svyLang", "English"), use.node = TRUE)
}
\description{
extract the choice labels of a multiple choice question in a language
}
