% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{node}
\alias{node}
\alias{node<-}
\title{get or set the node of a \code{svq|svy}}
\usage{
node(x)

node(x) <- value
}
\description{
get or set the node of a \code{svq|svy}
}
\details{
the node is the metadata corresponding to a `svy` or `svq` as
produced by `jsonlite::fromJSON` on the ODK form in `json` format.  it
is a nested list of character vectors containging information such as the
name, question type, labels in various languages, skip logic, constraints, 
etc.
}
