% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{preserve}
\alias{preserve}
\title{perform an arbitrary function on an object, preserving some attributes}
\usage{
preserve(x, ...)
}
\description{
perform an arbitrary function on an object, preserving some attributes
}
