% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservingly.R
\name{preservingly}
\alias{preservingly}
\title{modify a function acting on a list to preserve some of the attributes of
the list and its elements}
\usage{
preservingly(.f, .names = getOption("preserve_atts", c("node", "group",
  "class", "languages")))
}
\arguments{
\item{.f}{the function to be wrapped}

\item{.names}{vector of names of attributes to be preserved}
}
\description{
modify a function acting on a list to preserve some of the attributes of
the list and its elements
}
\details{
this adverb wraps destructive functions such as 
\code{dplyr::col_bind}
and all the row subsetting operations (e.g. \code{[]}, \code{[]<-}) 
so that \code{svy} 
objects and their constituent \code{svq}s do not lose their attributes
}
\note{
do not wrap functions that do not return a list of the same length,
i.e. column operations should be performed separately from row operations
}
