% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protectively.R
\name{protectively}
\alias{protectively}
\title{modify a function acting on a list to preserve matrix elements and 
some of the attributes of the list and its elements}
\usage{
protectively(.f, .names = getOption("preserve_atts", c("node", "group",
  "class", "languages", "groups")), recurse = TRUE)
}
\arguments{
\item{.f}{the function to be wrapped}

\item{.names}{vector of names of attributes to be preserved}
}
\description{
modify a function acting on a list to preserve matrix elements and 
some of the attributes of the list and its elements
}
\details{
this adverb wraps attribute-destructive functions such as 
\code{dplyr::col_bind}
and all the row subsetting operations (e.g. \code{[]}, \code{[]<-}) 
so that \code{svy} 
objects and their constituent \code{svq}s do not lose their attributes
TODO: handle grouped tibbles.
}
\note{
do not wrap functions that do not return a list of the same length,
i.e. column operations should be performed separately from row operations.
}
