% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{push.svy}
\alias{push.svy}
\title{push a svy object up to the database}
\usage{
\method{push}{svy}(s, name = getOption("svyDBSchema"), prefix = NULL,
  indexes = NULL, overwrite = FALSE, con = connection(), ...)
}
\arguments{
\item{s}{a \code{svy} object}

\item{name}{cannonical, unique survey name.  by convention:
<project-id>_<project-name>[__<survey-name>].}

\item{indexes}{ignored for now}

\item{overwrite}{whether to overwrite existing tables}

\item{con}{a \code{PqConnection} to the survey database

Push data and metadata of the survey up to the server.  The schema  
takes the name of the survey and contains three tables: 
 * instance, containing the collected data with one column per question 
 * question, containing the form structure, including group information, 
 name information, node, type and labels, one row per question
 * choice, containing the choices for multiple choice questions, including
 question, name, and labels, one row per question choice.
 
 @return TRUE for success, FALSE otherwise
 
 @export}
}
\description{
push a svy object up to the database
}
