% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{doSQL}
\alias{doSQL}
\alias{getSQL}
\title{run an sql send/get query}
\usage{
doSQL(..., con = connection())

getSQL(..., con = connection())
}
\arguments{
\item{...}{optional parameters to pass on to \code{db(Send|Get)query}}

\item{statement}{a string containing a valid SQL command}
}
\description{
run an sql send/get query
}
