% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{view}
\alias{view}
\title{show a dataframe as html in the viewer so you can copy and paste into a
spreadsheet}
\usage{
view(df)
}
\description{
show a dataframe as html in the viewer so you can copy and paste into a
spreadsheet
}
