% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform.R
\name{xlsform}
\alias{xlsform}
\title{return a list of data.frames corresponding to the three tabs of an xlsform}
\usage{
xlsform(form)
}
\description{
return a list of data.frames corresponding to the three tabs of an xlsform
}
