% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy.R
\name{as_tibble.odk_data}
\alias{as_tibble.odk_data}
\title{S3 generic for converting odk data in it's native json format to a tibble}
\usage{
\method{as_tibble}{odk_data}(d)
}
\description{
S3 generic for converting odk data in it's native json format to a tibble
}
