% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{label.svq}
\alias{label.svq}
\title{S3 method for extracting a \code{svq}'s label}
\usage{
\method{label}{svq}(x, lang = if (is.null(languages(x))) NULL else
  getOption("svyLang", "English"), use.node = TRUE)
}
\description{
S3 method for extracting a \code{svq}'s label
}
