% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{push.svy}
\alias{push.svy}
\title{push a svy object up to the database}
\usage{
\method{push}{svy}(s, name = deparse(substitute(s)),
  schema = getOption("svyDBSchema"), indexes = NULL,
  overwrite = FALSE, ...)
}
\description{
push a svy object up to the database
}
