% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{push.svy}
\alias{push.svy}
\title{push a svy object up to the database}
\usage{
\method{push}{svy}(s, name = getOption("svyName",
  getOption("svyProject")), prefix = NULL, indexes = NULL,
  overwrite = FALSE, con = connection(), ...)
}
\arguments{
\item{s}{a \code{svy} object}

\item{name}{cannonical, unique survey name.  by convention:
<project-id>_<project-name>[_<survey-name>].}
}
\description{
push a svy object up to the database
}
