% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push.svy}
\alias{push.svy}
\title{push a svy object up to the database}
\usage{
\method{push}{svy}(s, indexes = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{s}{a \code{svy} object}

\item{indexes}{ignored for now}

\item{overwrite}{whether to overwrite existing tables}

\item{name}{cannonical, unique survey name.  by convention:
<project-id>_<project-name>[__<survey-name>].}

\item{con}{a \code{PqConnection} to the survey database

Push data and metadata of the survey up to the server.  The schema  
takes the name of the survey and contains at least five tables: 
 * instance, containing the collected data with one column per question 
 * question, containing the form structure, including group information, 
 name information, node, type and labels, one row per question
 * choice, containing the choices for multiple choice questions, including
 question and the choice name
 * label, containing the text of all question labels in all 
 available languages, with question, language and text columns
 * choice_label, containing the test of all choice labels, with question,
 choice, language and text columns
 
 @return a list of tibbles connected to the database
 
 @export}
}
\description{
push a svy object up to the database
}
