% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{setSchema}
\alias{setSchema}
\alias{getSchema}
\title{set/get the current default schema of a database connection}
\usage{
setSchema(name = getOption("svyDBSchema", stop("need a schema")),
  con = connection())

getSchema()
}
\arguments{
\item{name}{string conntaining the schema}
}
\description{
set/get the current default schema of a database connection
}
