% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy.R
\name{svy}
\alias{svy}
\title{create a svy object}
\usage{
svy(dat = kobo_data(), form = kobo_form(), group = NULL)
}
\arguments{
\item{dat}{a tibble containing the raw data as returned by `as_tibble()`
on odk data as stored in ODK's native json format}

\item{form}{a list created using [jsonlite::fromJSON()] on ODK's native json
format.}

\item{group}{this is only used when called on a roster within a group.}
}
\description{
`svy()` is a [tibble()] with one element for each question containing the
answers and all metadata in the attributes.
}
\examples{
use the default functions to retrieve the data from a kobo server
\dontrun{
s <- svy()
}
load a svy from local data:
\dontrun{
s <- svy(node = jsonlite::fromJSON("data/form.json"),
dat = jsonlite::fromJSON("data/data.json"))
}
}
